fix(date-time-picker): update demo
