+++ src/date-time-picker/demos/time.vue
@@ -1,11 +1,11 @@
 <template>
-  <t-cell title="选择日期(年月日)" :note="pickerValue || '年 月 日'" @click="visible = true" />
+  <t-cell title="选择日期(时分)" :note="value || '时 分'" @click="visible = true" />
   <t-popup v-model="visible" placement="bottom">
     <t-date-time-picker
-      :value="pickerValue"
-      :mode="['date']"
+      v-model="value"
+      :mode="[null, 'minute']"
       title="选择日期"
-      format="YYYY-MM-DD"
+      format="HH:mm"
       @change="onChange"
       @pick="onPick"
       @confirm="onConfirm"
@@ -16,15 +16,16 @@
 <script lang="ts" setup>
 import { ref } from 'vue';
 import { cell } from 'tdesign-mobile-vue';
+import { DateValue } from '../type';
 
 const visible = ref(false);
-const pickerValue = ref('');
+const value = ref('');
 
-const onChange = (value: string) => {
+const onChange = (value: DateValue) => {
   console.log('change: ', value);
 };
 
-const onPick = (value: string) => {
+const onPick = (value: DateValue) => {
   console.log('pick: ', value);
 };
 
@@ -33,9 +34,8 @@ const onCancel = () => {
   visible.value = false;
 };
 
-const onConfirm = (value: string) => {
+const onConfirm = (value: DateValue) => {
   console.log('confirm: ', value);
-  pickerValue.value = value;
   visible.value = false;
 };
 </script>
